src/main/java/ru/javawebinar/graduation/web/vote/AdminVoteController.java
package ru.javawebinar.graduation.web.vote;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.*;
import ru.javawebinar.graduation.service.VoteService;
import ru.javawebinar.graduation.to.VoteTo;

import java.time.LocalDate;
import java.util.List;

@RestController
@RequestMapping(value = AdminVoteController.REST_URL, produces = MediaType.APPLICATION_JSON_VALUE)
public class AdminVoteController {
    static final String REST_URL = "/rest/admin/users";

    private final Logger log = LoggerFactory.getLogger(getClass());

    @Autowired
    private VoteService service;

    @GetMapping(value = "/votes")
    @ResponseStatus(HttpStatus.OK)
    public List<VoteTo> getToday() {
        log.info("get today Vote for all users");
        return service.getBetweenDates(LocalDate.now(), LocalDate.now());
    }

}
